+++ mne/viz/tests/test_utils.py
@@ -8,7 +8,8 @@ import numpy as np
 from nose.tools import assert_true, assert_raises
 from numpy.testing import assert_allclose
 
-from mne.viz.utils import compare_fiff, _fake_click, _compute_scalings
+from mne.viz.utils import (compare_fiff, _fake_click, _compute_scalings,
+                           _validate_if_list_of_axes)
 from mne.viz import ClickableImage, add_background_image, mne_analyze_colormap
 from mne.utils import run_tests_if_main
 from mne.io import read_raw_fif
@@ -115,6 +116,25 @@ def test_auto_scale():
     epochs.pick_types(eeg=True, meg=False)
     assert_raises(ValueError, _compute_scalings,
                   dict(grad='auto'), epochs)
+
+
+def test_validate_if_list_of_axes():
+    import matplotlib.pyplot as plt
+    fig, ax = plt.subplots(2, 2)
+    assert_raises(ValueError, _validate_if_list_of_axes, ax)
+    ax_flat = ax.ravel()
+    ax = ax.ravel().tolist()
+    _validate_if_list_of_axes(ax_flat)
+    _validate_if_list_of_axes(ax_flat, 4)
+    assert_raises(ValueError, _validate_if_list_of_axes, ax_flat, 5)
+    assert_raises(ValueError, _validate_if_list_of_axes, ax, 3)
+    assert_raises(ValueError, _validate_if_list_of_axes, 'error')
+    assert_raises(ValueError, _validate_if_list_of_axes, ['error'] * 2)
+    assert_raises(ValueError, _validate_if_list_of_axes, ax[0])
+    assert_raises(ValueError, _validate_if_list_of_axes, ax, 3)
+    ax_flat[2] = 23
+    assert_raises(ValueError, _validate_if_list_of_axes, ax_flat)
+    _validate_if_list_of_axes(ax, 4)
 
 
 run_tests_if_main()